vscode-container-client: align test file layout and hoist resolveCreatedAt
Rename/relocate unit test files so each mirrors its source: foo.ts -> foo.test.ts under a src/test subtree that matches the source's directory. Merges the individualized Shared* tests back onto their source names (SharedInspectContainerRecord.test.ts, SharedListContainerRecord.test.ts) and the DockerClientBase parsing tests onto DockerClientBase.test.ts.

Extract resolveCreatedAt into its own module and add resolveCreatedAtBaseline, reused by the inspect normalizer for its created/started/finished baseline instead of duplicating the parse-with-now-fallback logic.


Copilot-Session: c3540c82-5d55-404c-82b6-48331dbc1455
