+++ packages/vscode-container-client/src/test/clients/DockerClientBase/SharedListContainerRecord.test.ts
@@ -5,7 +5,7 @@
 
 import { expect } from 'chai';
 import { describe, it } from 'mocha';
-import { normalizeContainerState } from '../clients/DockerClientBase/SharedListContainerRecord';
+import { normalizeContainerState } from '../../../clients/DockerClientBase/SharedListContainerRecord';
 
 describe('(unit) normalizeContainerState', () => {
