+++ packages/vscode-container-client/src/test/clients/DockerClientBase/tryParseSize.test.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { expect } from 'chai';
-import { tryParseSize } from '../clients/DockerClientBase/tryParseSize';
+import { tryParseSize } from '../../../clients/DockerClientBase/tryParseSize';
 
 const KB = 1024;
 const MB = KB * 1024;
