+++ packages/vscode-container-client/src/test/utils/parseDockerLikeImageName.test.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { expect } from 'chai';
-import { parseDockerLikeImageName } from '../utils/parseDockerLikeImageName';
+import { parseDockerLikeImageName } from '../../utils/parseDockerLikeImageName';
 
 describe('(unit) parseDockerLikeImageName', () => {
     describe('Should parse valid image names', () => {
